Refactor core.
