--- js/.gitignore
@@ -1,1 +1,0 @@
-all*.js
