--- js/ui/.gitignore
@@ -1,1 +1,0 @@
-ui-all*.js
